refactor, add record processor
